bfp-simulation-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
stable-baselines3[extra]
gymnasium
opencv-python-headless
numpy
Pillow
